Fix passing the config to the command as stdin
The remaining config file is passed into the executed command as stdin,
because exec is called while the config files is fed into the read
command in the while loop. This causes misbehaviour in the executed
command.

Move the exec call out of the while loop and put it at the end of the
script. This ensures that the config file is closed.